# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ultra Venues API Docs",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/venues": {
      "get": {
        "tags": [
          "Venue"
        ],
        "summary": "List all venues",
        "description": "This endpoint retrieves a list of all venues, excluding sensitive information like password and isActive status.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/venues/{id}": {
      "get": {
        "tags": [
          "Venue"
        ],
        "summary": "Get venue details",
        "description": "This endpoint retrieves details of a specific venue by its ID, excluding sensitive information like password and isActive status.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the venue to retrieve details for"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Venue"
        ],
        "summary": "Update venue details",
        "description": "This endpoint updates the details of a specific venue by its ID, including uploading new pictures.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the venue to update"
          },
          {
            "name": "name",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "mobile": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                },
                "website": {
                  "example": "any"
                },
                "services": {
                  "example": "any"
                }
              }
            },
            "description": "Name of the venue"
          },
          {
            "name": "type",
            "in": "body",
            "description": "Type of the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          {
            "name": "email",
            "in": "body",
            "description": "Email of the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          {
            "name": "mobile",
            "in": "body",
            "description": "Mobile number of the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          {
            "name": "location",
            "in": "body",
            "description": "Location of the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          {
            "name": "website",
            "in": "body",
            "description": "Website URL of the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          {
            "name": "services",
            "in": "body",
            "description": "Services provided by the venue",
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/venues/update-email/{id}": {
      "post": {
        "tags": [
          "Venue"
        ],
        "summary": "Update venue email",
        "description": "This endpoint updates the email address of a specific venue by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the venue to update email for"
          },
          {
            "name": "newEmail",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "newEmail": {
                  "example": "any"
                }
              }
            },
            "description": "New email address of the venue",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/vendor/register": {
      "post": {
        "tags": [
          "Vendor"
        ],
        "summary": "Register a new venue",
        "description": "This endpoint allows a new venue to register by providing details such as venue type, name, address, location, postal code, description, facilities, phone number, email, website URL, and password.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Venue registration details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "venueType": {
                  "type": "string",
                  "example": "Conference Hall"
                },
                "venueName": {
                  "type": "string",
                  "example": "Grand Conference Center"
                },
                "address": {
                  "type": "string",
                  "example": "123 Main St"
                },
                "location": {
                  "type": "string",
                  "example": "Cityville"
                },
                "postalCode": {
                  "type": "string",
                  "example": "12345"
                },
                "description": {
                  "type": "string",
                  "example": "A spacious conference hall"
                },
                "facilities": {
                  "type": "string",
                  "example": "WiFi, Parking, AV Equipment"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "1234567890"
                },
                "email": {
                  "type": "string",
                  "example": "venue@example.com"
                },
                "websiteUrl": {
                  "type": "string",
                  "example": "http://venue.com"
                },
                "password": {
                  "type": "string",
                  "example": "securepassword"
                }
              },
              "required": [
                "venueType",
                "venueName",
                "address",
                "location",
                "postalCode",
                "description",
                "facilities",
                "phoneNumber",
                "email",
                "websiteUrl",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/vendor/login": {
      "post": {
        "tags": [
          "Vendor"
        ],
        "summary": "Venue login",
        "description": "This endpoint allows a venue to login by providing their email and password.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Venue login details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "venue@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "securepassword"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/vendor/dashboard/{venueEmail}": {
      "get": {
        "tags": [
          "Vendor"
        ],
        "summary": "Fetch venue dashboard data",
        "description": "This endpoint provides venue details and statistics, including total revenue, total visitors, upcoming bookings, and visitor counts by month.",
        "parameters": [
          {
            "name": "venueEmail",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Email address of the venue"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/vendor/transactions/{email}": {
      "get": {
        "tags": [
          "Vendor"
        ],
        "summary": "Fetch transactions for a specific email",
        "description": "This endpoint provides all transactions for a specific email, including total revenue, current month revenue, and previous month revenue.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Email address of the venue"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}